+++ packages/zowe-explorer/CHANGELOG.md
@@ -17,6 +17,7 @@ All notable changes to the "vscode-extension-for-zowe" extension will be documen
 - Added Display confirmation dialog when submitting local JCL. [#2061](https://github.com/zowe/vscode-extension-for-zowe/issues/2061)
 - Added support for adding a Zowe profile across all trees [#2603](https://github.com/zowe/vscode-extension-for-zowe/issues/2603)
 - Added "Filter Jobs" feature in Jobs tree view: accessible via filter icon or right-clicking on session node. [#2599](https://github.com/zowe/vscode-extension-for-zowe/issues/2599)
+- Added z/OS System Name (SMFID) to Zowe Explorer Jobs View. [#2629] (https://github.com/zowe/vscode-extension-for-zowe/pull/2629)
 
 ### Bug fixes